Mapeamento feito com sucesso

+++ Calango/ViacaoCalango/ViacaoCalango/Controllers/ReservaController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using ViacaoCalango.Application.AppService;
+using ViacaoCalango.Application.Interfaces;
 using ViacaoCalango.Application.ViewModels;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -12,12 +13,32 @@ namespace ViacaoCalango.API.Controllers
     [ApiController]
     public class ReservaController : ControllerBase
     {
+        private readonly IReservaAppService _reservaAppService;
+
+        public ReservaController(IReservaAppService reservaAppService)
+        {
+            _reservaAppService = reservaAppService;
+        }
+
+        public ReservaController() { }
+
+
         // GET: api/<ReservaController>
         [HttpGet]
         public List<string> Get()
         {
-            ReservaAppService reservaAppService = new ReservaAppService();
-            var listaReservas = reservaAppService.GetAll();
+            List<ReservaViewModel> listaReservas = null;
+
+            if (_reservaAppService == null)
+            {
+                ReservaAppService reservaAppService = new ReservaAppService();
+                listaReservas = reservaAppService.GetAll();
+            }
+            else
+            {
+                listaReservas = _reservaAppService.GetAll();
+            }
+
             return listaReservas.Select(s => s.toString()).ToList();
         }
 
@@ -25,41 +46,61 @@ namespace ViacaoCalango.API.Controllers
         [HttpGet("{id}")]
         public string Get(int id)
         {
-            ReservaAppService reservaAppService = new ReservaAppService();
-            return reservaAppService.GetById(id).toString();
+            if(_reservaAppService == null)
+            {
+                 ReservaAppService reservaAppService = new ReservaAppService();
+                 return reservaAppService.GetById(id).toString();
+            }
+
+            return "Erro";
+           
         }
 
         // POST api/<ReservaController>
         [HttpPost]
         public void Post([FromBody] ReservaViewModel reservaViewModel)
         {
-            ReservaAppService reservaAppService = new ReservaAppService();
-            if(reservaAppService.Create(reservaViewModel))
+            if (_reservaAppService == null)
             {
-                StatusCode(201);
+                ReservaAppService reservaAppService = new ReservaAppService();
+                if(reservaAppService.Create(reservaViewModel))
+                {
+                    StatusCode(201);
+                }
             }
+            StatusCode(203);
         }
 
         // PUT api/<ReservaController>/5
         [HttpPut("{id}")]
         public void Put(int id, [FromBody] ReservaViewModel reservaViewModel)
         {
-            ReservaAppService reservaAppService = new ReservaAppService();
-            if(reservaAppService.Update(id, reservaViewModel))
+            if (_reservaAppService == null)
             {
-                StatusCode(202);
+                ReservaAppService reservaAppService = new ReservaAppService();
+                if(reservaAppService.Update(id, reservaViewModel))
+                {
+                    StatusCode(202);
+                }
             }
+
+            StatusCode(404);
+            
         }
 
         // DELETE api/<ReservaController>/5
         [HttpDelete("{id}")]
         public void Delete(int id)
         {
-            ReservaAppService reservaAppService = new ReservaAppService();
-            if(reservaAppService.Delete(id))
+            if(_reservaAppService == null)
             {
-                StatusCode(203);
+                ReservaAppService reservaAppService = new ReservaAppService();
+                if(reservaAppService.Delete(id))
+                {
+                    StatusCode(203);
+                }
             }
+            StatusCode(404);
         }
     }
 }
